uhack-backend/requirements.txt

fastapi
uvicorn
python-multipart
openai-whisper
numpy
opencv-python
